feat: Automatically convert cty Tubles to cty Lists when needed
Specifically, when cty is decoding json, and the top json level is an
array, it decodes it into a Tuble (as the inner elements could be
different).

This patch trys to convert it to a cty List, if and only if the target
type is a list or a slice AND the elements in the Tuble are all of same
type.